Extract protocol to separated module and separated tag
