--- generate.go
@@ -1,3 +1,0 @@
-//go:generate sh bin/compile_proto.sh ./proto ./proto proto/*.proto
-
-package main
