+++ go.mod
@@ -2,13 +2,16 @@ module github.com/goverland-labs/goverland-datasource-snapshot
 
 go 1.21
 
+replace github.com/goverland-labs/goverland-datasource-snapshot/protocol => ./protocol
+
 require (
 	github.com/Yamashou/gqlgenc v0.14.0
 	github.com/akamensky/argparse v1.4.0
 	github.com/caarlos0/env/v8 v8.0.0
-	github.com/ethereum/go-ethereum v1.13.5
-	github.com/google/uuid v1.3.1
+	github.com/ethereum/go-ethereum v1.13.14
+	github.com/google/uuid v1.6.0
 	github.com/gorilla/mux v1.8.0
+	github.com/goverland-labs/goverland-datasource-snapshot/protocol v0.0.0
 	github.com/goverland-labs/goverland-platform-events v0.2.1
 	github.com/goverland-labs/snapshot-sdk-go v0.4.0
 	github.com/grpc-ecosystem/go-grpc-middleware/v2 v2.0.1
@@ -22,8 +25,7 @@ require (
 	github.com/shutter-network/shutter/shlib v0.1.11
 	github.com/smartystreets/goconvey v1.8.0
 	github.com/vektah/gqlparser/v2 v2.5.3
-	google.golang.org/grpc v1.59.0
-	google.golang.org/protobuf v1.31.0
+	google.golang.org/grpc v1.62.0
 	gorm.io/driver/postgres v1.5.2
 	gorm.io/gorm v1.25.2
 )
@@ -34,6 +36,7 @@ require (
 	github.com/cespare/xxhash/v2 v2.2.0 // indirect
 	github.com/golang/protobuf v1.5.3 // indirect
 	github.com/gopherjs/gopherjs v1.17.2 // indirect
+	github.com/holiman/uint256 v1.2.4 // indirect
 	github.com/jackc/pgpassfile v1.0.0 // indirect
 	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
 	github.com/jackc/pgx/v5 v5.3.1 // indirect
@@ -50,10 +53,11 @@ require (
 	github.com/prometheus/common v0.45.0 // indirect
 	github.com/prometheus/procfs v0.12.0 // indirect
 	github.com/smartystreets/assertions v1.13.1 // indirect
-	golang.org/x/crypto v0.14.0 // indirect
-	golang.org/x/exp v0.0.0-20230905200255-921286631fa9 // indirect
-	golang.org/x/net v0.17.0 // indirect
-	golang.org/x/sys v0.15.0 // indirect
-	golang.org/x/text v0.13.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20230822172742-b8732ec3820d // indirect
+	golang.org/x/crypto v0.18.0 // indirect
+	golang.org/x/exp v0.0.0-20231110203233-9a3e6036ecaa // indirect
+	golang.org/x/net v0.20.0 // indirect
+	golang.org/x/sys v0.16.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240123012728-ef4313101c80 // indirect
+	google.golang.org/protobuf v1.32.0 // indirect
 )
