+++ internal/voting/grpc_server.go
@@ -8,7 +8,7 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	"github.com/goverland-labs/goverland-datasource-snapshot/proto/votingpb"
+	"github.com/goverland-labs/goverland-datasource-snapshot/protocol/votingpb"
 )
 
 type GrpcServer struct {
