+++ protocol/generate.go
@@ -1,0 +1,3 @@
+//go:generate sh bin/compile_proto.sh ./ ./ */*.proto
+
+package protocol
