+++ protocol/votingpb/voting.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-package voting;
+package votingpb;
 
 import "google/protobuf/any.proto";
